Compatibility for MS Webfolders Client
Microsoft Webfolders need the createddate and getlastmodified in
specific, obsolete formats.  See
http://www.greenbytes.de/tech/webdav/webfolder-client-list.html#issue-date-format

This sends the date in an rfc1123 format (as required by spec) as well
as MS-compatible format. Previous author had commented out code to send
it in the rfc1123 format...but it didn't work, presumably because of the
missing uuid that microsoft requires.